--- jetson/odrive_bridge/odrive_bridge/project.ini
@@ -1,4 +1,0 @@
-[build]
-lang=python
-executable=True
-deps=rover_msgs
