+++ jetson/odrive_bridge/src/__main__.py
@@ -33,6 +33,11 @@ def main():
     global lock
     global speedlock
 
+    global start_time
+    global watchdog
+
+    start_time = t.clock()
+
     legal_controller = int(sys.argv[1])
     legal_axis = sys.argv[2]
 
@@ -49,10 +54,16 @@ def main():
     # start up sequence is called, disconnected-->disarm-->arm
 
     while True:
+        watchdog = t.clock() - start_time
+        if (watchdog > 1.0):
+            print("loss of comms")
+            left_speed = 0
+            right_speed = 0
+
         try:
             odrive_bridge.update()
-
         except fibre.protocol.ChannelBrokenException:
+            print("odrive has been unplugged")
             lock.acquire()
             odrive_bridge.on_event("disconnected odrive")
             lock.release()
@@ -66,12 +77,22 @@ def lcmThreaderMan():
     lcm_1.subscribe("/drive_vel_cmd", drive_vel_cmd_callback)
     while True:
         lcm_1.handle()
+        global start_time
+        start_time = t.clock()
+        try:
+            publish_encoder_msg()
+        except NameError:
+            pass
+        except AttributeError:
+            pass
+        except fibre.protocol.ChannelBrokenException:
+            pass
 
 
-events = ["disconnected odrive", "disarm cmd", "arm cmd", "calibrate cmd", "odrive error"]
-states = ["DisconnectedState", "DisarmedState", "ArmedState", "ErrorState"]
-# Program states possible - BOOT,  DISARMED, ARMED, ERROR
-# 							1		 2	      3	      4
+events = ["disconnected odrive", "disarm cmd", "arm cmd", "calibrating cmd", "odrive error"]
+states = ["DisconnectedState", "DisarmedState", "ArmedState", "CalibrateState", "ErrorState"]
+# Program states possible - BOOT,  DISARMED, ARMED, CALIBRATE ERROR
+# 							1		 2	      3	       4        5
 
 
 class State(object):
@@ -133,9 +154,6 @@ class DisarmedState(State):
 
         elif (event == "calibrating cmd"):
             # sequence can be moved to armed ?
-            modrive.calibrate()
-            print("done calibrating")
-            modrive.disarm()
             return DisarmedState()
 
         elif (event == "odrive error"):
@@ -160,6 +178,25 @@ class ArmedState(State):
 
         elif (event == "odrive error"):
             return ErrorState()
+
+        elif (event == "calibrating cmd"):
+            modrive.reset()
+            return CalibrateState()
+
+        return self
+
+
+class CalibrateState(State):
+    def on_event(self, event):
+        global modrive
+
+        if (event == "arm cmd"):
+            modrive.arm()
+            return ArmedState()
+
+        if(modrive.check_errors()):
+            print("clearing calibration errors")
+            dump_errors(modrive.odrive, True)
 
         return self
 
@@ -197,7 +234,12 @@ class OdriveBridge(object):
         global modrive
         global legal_controller
         print("looking for odrive")
-        odrives = ["2091358E524B", "20563591524B"]
+
+        # odrive 0 --> front motors
+        # odrive 1 --> middle motors
+        # odrive 2 --> back motors
+
+        odrives = ["205F3883304E", "2091358E524B", "2084399C4D4D"]
         id = odrives[legal_controller]
 
         print(id)
@@ -213,7 +255,7 @@ class OdriveBridge(object):
         Incoming events are
         delegated to the given states which then handle the event.
         The result is then assigned as the new state.
-        The events we can send are disarm cmd, arm cmd, and calibrate cmd.
+        The events we can send are disarm cmd, arm cmd, and calibrating cmd.
         """
 
         print("on event called, event:", event)
@@ -222,16 +264,10 @@ class OdriveBridge(object):
         publish_state_msg(state_msg, odrive_bridge.get_state())
 
     def update(self):
-        if (str(self.state) == "DisarmedState"):
-            if (t.time() - self.encoder_time > 0.1):  # order is flipped? why?
-                self.encoder_time = publish_encoder_msg(vel_msg)
-
-        elif (str(self.state) == "ArmedState"):
+        if (str(self.state) == "ArmedState"):
             global speedlock
             global left_speed
             global right_speed
-            if (self.encoder_time - t.time() > 0.1):
-                self.encoder_time = publish_encoder_msg(vel_msg)
 
             speedlock.acquire()
             modrive.set_vel("LEFT", left_speed)
@@ -244,9 +280,21 @@ class OdriveBridge(object):
             self.on_event("arm cmd")
             lock.release()
 
+        elif (str(self.state) == "CalibrateState"):
+            self.connect()
+            modrive.calibrate()
+            self.connect()
+            print("done calibrating")
+
+            lock.acquire()
+            self.on_event("arm cmd")
+            lock.release()
+
         errors = modrive.check_errors()
 
         if errors:
+            # if (errors == 0x800 or erros == 0x1000):
+
             lock.acquire()
             self.on_event("odrive error")
             lock.release()
@@ -271,23 +319,25 @@ def publish_state_msg(msg, state):
     print("changed state to " + state)
 
 
-def publish_encoder_helper(msg, axis):
+def publish_encoder_helper(axis):
     global modrive
     global legal_controller
+    msg = DriveVelData()
     msg.measuredCurrent = modrive.get_iq_measured(axis)
     msg.estimatedVel = modrive.get_vel_estimate(axis)
 
     motor_map = {("LEFT", 0): 0, ("RIGHT", 0): 1,
-                 ("LEFT", 1): 2, ("RIGHT", 1): 3}
+                 ("LEFT", 1): 2, ("RIGHT", 1): 3,
+                 ("LEFT", 2): 4, ("RIGHT", 2): 5}
+
     msg.axis = motor_map[(axis, legal_controller)]
 
     lcm_.publish("/drive_vel_data", msg.encode())
 
 
-def publish_encoder_msg(msg):
-    publish_encoder_helper(msg, "LEFT")
-    publish_encoder_helper(msg, "RIGHT")
-    return t.time()
+def publish_encoder_msg():
+    publish_encoder_helper("LEFT")
+    publish_encoder_helper("RIGHT")
 
 
 def drive_state_cmd_callback(channel, msg):
@@ -320,7 +370,7 @@ def drive_vel_cmd_callback(channel, msg):
             right_speed = cmd.right
             speedlock.release()
     except NameError:
-        print("waiting to find odrive")
+        pass
 
 
 if __name__ == "__main__":
@@ -343,27 +393,27 @@ class Modrive:
             return getattr(self, attr)
         return getattr(self.odrive, attr)
 
-    def calibrate(self):
+    def reset(self):
         self._reset(self.front_axis)
         self._reset(self.back_axis)
         self.odrive.save_configuration()
         # the guide says to reboot here...
 
-        self._requested_state("LEFT", AXIS_STATE_FULL_CALIBRATION_SEQUENCE)
-        while (self.get_current_state("RIGHT") != AXIS_STATE_IDLE):
-            pass
-        self._requested_state("LEFT", AXIS_STATE_FULL_CALIBRATION_SEQUENCE)
-        while (self.get_current_state("RIGHT") != AXIS_STATE_IDLE):
-            pass
+    def calibrate(self):
+        dump_errors(self.odrive, True)  # clears all odrive encoder errors
+        self._requested_state(AXIS_STATE_FULL_CALIBRATION_SEQUENCE)
 
         front_state, back_state = self.get_current_state()
 
-        # if both are idle it means its done calibrating
-        if front_state == AXIS_STATE_IDLE \
-                and back_state == AXIS_STATE_IDLE:
-            self._pre_calibrate(self.front_axis)
-            self._pre_calibrate(self.back_axis)
-            self.odrive.save_configuration()
+        # if both axes are idle it means its done calibrating
+        while(front_state != AXIS_STATE_IDLE
+                or back_state != AXIS_STATE_IDLE):
+            front_state, back_state = self.get_current_state()
+            pass
+
+        self._pre_calibrate(self.front_axis)
+        self._pre_calibrate(self.back_axis)
+        self.odrive.save_configuration()
         # also says to reboot here...
 
     def disarm(self):
@@ -392,6 +442,7 @@ class Modrive:
         self._set_control_mode(CTRL_MODE_VELOCITY_CONTROL)
 
     def get_iq_measured(self, axis):
+        # measured current [Amps]
         if (axis == "LEFT"):
             return self.front_axis.motor.current_control.Iq_measured
         elif(axis == "RIGHT"):
@@ -416,9 +467,9 @@ class Modrive:
 
     def set_vel(self, axis, vel):
         if (axis == "LEFT"):
-            self.front_axis.controller.vel_setpoint = vel * 300
+            self.front_axis.controller.vel_setpoint = vel * 205
         elif axis == "RIGHT":
-            self.back_axis.controller.vel_setpoint = vel * -300
+            self.back_axis.controller.vel_setpoint = vel * -205
 
     def get_current_state(self):
         return (self.front_axis.current_state, self.back_axis.current_state)
@@ -439,8 +490,8 @@ class Modrive:
         m_axis.controller.config.control_mode = CTRL_MODE_VELOCITY_CONTROL
 
     def _pre_calibrate(self, m_axis):
-        self.m_axis.motor.config.pre_calibrated = True
-        self.m_axis.encoder.config.pre_calibrated = True
+        m_axis.motor.config.pre_calibrated = True
+        m_axis.encoder.config.pre_calibrated = True
 
     def check_errors(self):
         front = self.front_axis.error
